feat: Refactoring of tools & types (#5)
* feat: Refactoring of tools & types

* Update package & build for publishing
