+++ package.json
@@ -1,9 +1,11 @@
 {
-  "name": "mcp-actualbudget",
+  "name": "actual-mcp",
   "version": "0.1.0",
   "description": "Actual Budget MCP server exposing API functionality",
   "private": false,
   "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "type": "module",
   "bin": {
     "actual-mcp": "build/index.js"
   },
@@ -50,5 +52,7 @@
     "ai",
     "model-context-protocol",
     "actual"
-  ]
+  ],
+  "author": "Stefan Stefanov",
+  "license": "MIT"
 }
