+++ src/core/aggregation/sum-by.ts
@@ -1,0 +1,2 @@
+// Generic sum-by utilities
+export {};
