+++ src/core/data/fetch-categories.ts
@@ -1,0 +1,10 @@
+import { getCategories, getCategoryGroups } from "../../actual-api.js";
+import type { Category, CategoryGroup } from "../../core/types/domain.js";
+
+export async function fetchAllCategories(): Promise<Category[]> {
+  return getCategories();
+}
+
+export async function fetchAllCategoryGroups(): Promise<CategoryGroup[]> {
+  return getCategoryGroups();
+}
