+++ src/core/index.ts
@@ -1,0 +1,13 @@
+// Core exports
+export * from "./input/argument-parser.js";
+export * from "./input/validators.js";
+export * from "./data/fetch-accounts.js";
+export * from "./data/fetch-categories.js";
+export * from "./data/fetch-transactions.js";
+export * from "./aggregation/group-by.js";
+export * from "./aggregation/sum-by.js";
+export * from "./aggregation/sort-by.js";
+export * from "./mapping/category-mapper.js";
+export * from "./mapping/transaction-mapper.js";
+export * from "./mapping/category-classifier.js";
+export * from "./types/domain.js";
