+++ src/core/input/argument-parser.ts
@@ -1,0 +1,2 @@
+// Shared argument parsing utilities will go here
+export {};
