+++ src/core/input/validators.ts
@@ -1,0 +1,2 @@
+// Shared validators for input data
+export {};
