+++ src/core/mapping/category-classifier.ts
@@ -1,0 +1,2 @@
+// Shared category classification logic
+export {};
