+++ src/core/mapping/transaction-mapper.ts
@@ -1,0 +1,2 @@
+// Shared transaction mapping logic
+export {};
