+++ src/core/types/domain.ts
@@ -1,0 +1,58 @@
+// Shared domain types and interfaces (Account, Transaction, Category, etc.)
+
+export interface Account {
+  id: string;
+  name: string;
+  type?: string;
+  offbudget?: boolean;
+  closed?: boolean;
+  balance?: number;
+}
+
+export interface Transaction {
+  id?: string;
+  account: string;
+  date: string;
+  amount: number;
+  payee?: string;
+  payee_name?: string;
+  category?: string;
+  category_name?: string;
+  notes?: string;
+}
+
+export interface Category {
+  id: string;
+  name: string;
+  group_id: string;
+  is_income?: boolean;
+}
+
+export interface CategoryGroup {
+  id: string;
+  name: string;
+  is_income?: boolean;
+}
+
+export interface CategoryGroupInfo {
+  id: string;
+  name: string;
+  isIncome: boolean;
+  isSavingsOrInvestment: boolean;
+}
+
+export interface CategorySpending {
+  id: string;
+  name: string;
+  group: string;
+  isIncome: boolean;
+  total: number;
+  transactions: number;
+}
+
+export interface GroupSpending {
+  name: string;
+  total: number;
+  categories: CategorySpending[];
+}
+
