+++ src/tools/balance-history/data-fetcher.ts
@@ -1,18 +1,27 @@
 // Fetches accounts, transactions, and balances for balance-history tool
-import { getAccounts, getTransactions } from "../../actual-api.js";
-import type { Account, Transaction } from "../../types.js";
+import { fetchAllAccounts } from "../../core/data/fetch-accounts.js";
+import { fetchTransactionsForAccount } from "../../core/data/fetch-transactions.js";
+import type { Account, Transaction } from "../../core/types/domain.js";
 import api from "@actual-app/api";
 
 export class BalanceHistoryDataFetcher {
-  async fetchAll(accountId: string, start: string, end: string): Promise<{
+  async fetchAll(
+    accountId: string,
+    start: string,
+    end: string
+  ): Promise<{
     accounts: Account[];
     account: Account | undefined;
     transactions: Transaction[];
     currentBalance: number;
   }> {
-    const accounts = await getAccounts();
+    const accounts = await fetchAllAccounts();
     const account = accounts.find((a) => a.id === accountId);
-    const transactions = await getTransactions(accountId, start, end);
+    const transactions = await fetchTransactionsForAccount(
+      accountId,
+      start,
+      end
+    );
     const currentBalance = await api.getAccountBalance(accountId);
     return { accounts, account, transactions, currentBalance };
   }
