+++ src/tools/spending-by-category/category-mapper.ts
@@ -1,53 +1,6 @@
 // Maps category IDs to names/groups and identifies income/savings/investment categories
-import type { Category, CategoryGroup } from "../../types.js";
+import type { Category, CategoryGroup } from "../../core/types/domain.js";
+import { CategoryMapper } from "../../core/mapping/category-mapper.js";
 
-export interface CategoryGroupInfo {
-  id: string;
-  name: string;
-  isIncome: boolean;
-  isSavingsOrInvestment: boolean;
-}
-
-export class SpendingByCategoryCategoryMapper {
-  categoryNames: Record<string, string> = {};
-  groupNames: Record<string, string> = {};
-  categoryToGroup: Record<string, CategoryGroupInfo> = {};
-  investmentCategories: Set<string> = new Set();
-
-  constructor(categories: Category[], categoryGroups: CategoryGroup[]) {
-    categories.forEach((cat) => {
-      this.categoryNames[cat.id] = cat.name;
-    });
-    categoryGroups.forEach((group) => {
-      this.groupNames[group.id] = group.name;
-    });
-    categories.forEach((cat) => {
-      const groupName = this.groupNames[cat.group_id] || "Unknown Group";
-      const isIncome = !!cat.is_income;
-      const isSavingsOrInvestment =
-        groupName.toLowerCase().includes("investment") ||
-        groupName.toLowerCase().includes("savings");
-      this.categoryToGroup[cat.id] = {
-        id: cat.group_id,
-        name: groupName,
-        isIncome,
-        isSavingsOrInvestment,
-      };
-      if (isSavingsOrInvestment) {
-        this.investmentCategories.add(cat.id);
-      }
-    });
-  }
-
-  getCategoryName(categoryId: string): string {
-    return this.categoryNames[categoryId] || "Unknown Category";
-  }
-
-  getGroupInfo(categoryId: string): CategoryGroupInfo | undefined {
-    return this.categoryToGroup[categoryId];
-  }
-
-  isInvestmentCategory(categoryId: string): boolean {
-    return this.investmentCategories.has(categoryId);
-  }
-}
+// This file now only re-exports CategoryMapper from core.
+export { CategoryMapper };
