+++ src/tools/spending-by-category/group-aggregator.ts
@@ -1,30 +1,6 @@
 // Aggregates category spendings into groups and sorts them
-import type { CategorySpending, GroupSpending } from "./types.js";
+import type { CategorySpending, GroupSpending } from "../../core/types/domain.js";
+import { GroupAggregator } from "../../core/aggregation/group-by.js";
 
-export class SpendingByCategoryGroupAggregator {
-  aggregateAndSort(
-    spendingByCategory: Record<string, CategorySpending>
-  ): GroupSpending[] {
-    const spendingByGroup: Record<string, GroupSpending> = {};
-    Object.values(spendingByCategory).forEach((category) => {
-      if (!spendingByGroup[category.group]) {
-        spendingByGroup[category.group] = {
-          name: category.group,
-          total: 0,
-          categories: [],
-        };
-      }
-      spendingByGroup[category.group].total += category.total;
-      spendingByGroup[category.group].categories.push(category);
-    });
-    // Sort groups by absolute total (descending)
-    const sortedGroups: GroupSpending[] = Object.values(spendingByGroup).sort(
-      (a, b) => Math.abs(b.total) - Math.abs(a.total)
-    );
-    // Sort categories within each group by absolute total (descending)
-    sortedGroups.forEach((group) => {
-      group.categories.sort((a, b) => Math.abs(b.total) - Math.abs(a.total));
-    });
-    return sortedGroups;
-  }
-}
+// This file now only re-exports GroupAggregator from core.
+export { GroupAggregator };
